gat: support podman

+++ gat/gat.py
@@ -570,7 +570,9 @@ class Conn:
         if self.dkTunnel is not None:
             dkRunUser = "-u %s" % self.dkId if self.dkId is not None else ""
             cmd = str2arg(cmd)
-            cmd = f'sudo docker exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
+            cmd = f'podman exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
+            if g_config.get("podman") != True:
+                cmd = f'sudo docker exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
             # alias defined in .bashrc is working but -l should be used for something in /etc/profile.d and .profile
             out = self.dkTunnel.ssh.runOutput(cmd, log=log)
         elif self.ssh is not None:
@@ -642,7 +644,9 @@ class Conn:
         if self.dkTunnel is not None:
             dkRunUser = "-u %s" % self.dkId if self.dkId is not None else ""
             cmd = str2arg(cmd)
-            cmd = f'docker exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
+            cmd = f'podman exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
+            if g_config.get("podman") != True:
+                cmd = f'docker exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
             return self.dkTunnel.ssh.runOutputAll(cmd)
         elif self.ssh is not None:
             return self.ssh.runOutputAll(cmd)
@@ -672,7 +676,9 @@ class Conn:
             dkRunUser = "-u %s" % self.dkId if self.dkId is not None else ""
             cmd = str2arg(cmd)
             # sudo docker로 하면 cmd에 '가 있으면 centos에서 실행이 안된다
-            cmd = f'docker exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
+            cmd = f'podman exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
+            if g_config.get("podman") != True:
+                cmd = f'docker exec -i {dkRunUser} {self.dkName} bash -c "{cmd}"'
             # print("run cmd(dk) - %s" % cmd)
             return self.dkTunnel.ssh.run(cmd)
         elif self.ssh is not None:
@@ -756,9 +762,11 @@ class Conn:
             # pp = f'/tmp/upload-{g_main.uid}.tmp'
             pp = f"{self.tempPathGet()}/upload.tmp"
             self.dkTunnel.ssh.uploadFile(src, pp)
-            self.dkTunnel.ssh.run(
-                f"sudo docker cp {pp} {self.dkName}:{dest} && rm -f {pp}"
-            )
+            cmd = f"podman cp {pp} {self.dkName}:{dest} && rm -f {pp}"
+            if g_config.get("podman") != True:
+                cmd = f"sudo docker cp {pp} {self.dkName}:{dest} && rm -f {pp}"
+
+            self.dkTunnel.ssh.run(cmd)
         elif self.ssh is None:
             self.run(f"cp {src} {dest}")
         else:
@@ -789,11 +797,11 @@ class Conn:
                         os.path.join("/tmp/gat_upload", pp[len(src) :], file),
                     )
 
-            self.run("rm -rf %s" % dest)
-            self.dkTunnel.ssh.run(
-                "sudo docker cp /tmp/gat_upload %s:%s && rm -rf /tmp/gat_upload"
-                % (self.dkName, dest)
-            )
+            self.run(f"rm -rf {dest}")
+            cmd = f"podman cp /tmp/gat_upload {self.dkName}:{dest} && rm -rf /tmp/gat_upload"
+            if g_config.get("podman") != True:
+                cmd = f"sudo docker cp /tmp/gat_upload {self.dkName}:{dest} && rm -rf /tmp/gat_upload"
+            self.dkTunnel.ssh.run(cmd)
         else:
             self.ssh.uploadFolder(src, dest)
 
@@ -813,9 +821,10 @@ class Conn:
                 if self.dkTunnel is not None:
                     pp3 = f"/tmp/gatHelperDk.py"
                     self.dkTunnel.uploadFile(src, pp3)
-                    self.dkTunnel.ssh.run(
-                        f"sudo docker cp {pp3} {self.dkName}:{pp2} && rm -f {pp3}"
-                    )
+                    cmd = f"podman cp {pp3} {self.dkName}:{pp2} && rm -f {pp3}"
+                    if g_config.get("podman") != True:
+                        cmd = f"sudo docker cp {pp3} {self.dkName}:{pp2} && rm -f {pp3}"
+                    self.dkTunnel.ssh.run(cmd)
                 else:
                     self.uploadFile(src, pp2)
